+++ core/computed/lcp-image-record.js
